Kerja_profesi-main/waiting_v2/requirements.txt
opencv-python
numpy
shapely
ultralytics
paho-mqtt
